\name{runBLCOPTests}
\alias{runBLCOPTests}

\title{Execute the BLCOP unit tests}
\description{
  Uses the \code{RUnit} package to execute a series of unit tests.
}
\usage{
runBLCOPTests(testPath = BLCOPOptions("unitTestPath"), protocolFile = "BLCOPTests.html", 
    writeProtocol = FALSE) 
}

\arguments{
  \item{testPath}{Location of the unit tests.}
  \item{protocolFile}{Name of the html report file generated by the RUnit function printHTMLProtocol}
  \item{writeProtocol}{Logical flag.  Should the above html report be produced?}
}

\value{
The summary of an object returned by RUnit's \code{runTestSuite}
}

\section{Warning}{
These unit tests are in need of additional test cases, and should not be regarded as exhaustive in their current state.
}


\author{Francisco Gochez <fgochez@mango-solutions.com> }

\examples{
\dontrun{
    runBLCOPTests()
    }
}
\keyword{misc}
\keyword{debugging}