\name{sp500Returns}
\alias{sp500Returns}
\docType{data}
\title{S\&P500 Returns}
\description{
    Monthly returns of the S&P 500 index for the period 2/2/1998 through  1/12/2003
}
\usage{sp500Returns}
\format{
    A matrix with 1 column and 71 rows.
}

\examples{
    ts.plot(sp500Returns)
}
\keyword{datasets}
           