\name{US13wTB}
\alias{US13wTB}
\docType{data}
\title{Risk free rate of return}
\description{
  The monthly rate of return of the US 13 week Treasury Bill for the period 30/1/1998 through 30/11/2003.
}
\usage{US13wTB}
\format{
    A one-column matrix with 71 rows.
}

\examples{
ts.plot(US13wTB)
}
\keyword{datasets}
